Script to rerun failed GitHub actions run (#5687)
# Motivation

Sometimes when a test fails because of a network issue or other
flakiness, we want to rerun the test.
Unfortunately, the rerun can only be triggered after all the other tests
have also finished.
For that it can be useful to keep trying to rerun in a loop until it
succeeds.

# Changes

Add `scripts/github-rerun` to which you can pass the URL of a run and it
will try to rerun that run once a minute until it reruns.

# Tests

Tried manually with
```
scripts/github-rerun https://github.com/dfinity/snsdemo/actions/runs/11514250845/job/32052488208
```
and
```
scripts/github-rerun https://github.com/dfinity/snsdemo/actions/runs/11514250846
```

# Todos

- [ ] Add entry to changelog (if necessary).
not necesssary