FOLLOW-244: Show a toast when neuron is refreshed (#5704)
# Motivation

When a neuron's account balance is out of sync with the neuron stake, we
refresh the neuron in the background when the user visits the neuron
detail page.
To make it clear what happened we want to show a toast to the user if
the neuron was refreshed.

# Changes

1. Show a toast if the neuron was refreshed in the background.

# Tests

1. Unit test updated.
2. Tested manually after disabling the refresh functionality in the
nns-dapp canister locally.

# Todos

- [ ] Add entry to changelog (if necessary).
covered by existing entry