FOLLOW-244: Stop refreshing neurons in nns-dapp canister (#5706)
# Motivation

In https://github.com/dfinity/nns-dapp/pull/5697 we added fallback
neuron refresh to the frontend in order to be able to remove it from the
backend.

This PR removes fallback neuron refresh from the nns-dapp backend
canister.

# Changes

1. Remove neuron top-up code from the backend canister.
2. Add cleanup TODOs for things that can be cleaned up after one or more
releases.

# Tests

1. Rust unit tests have been removed and updated.
2. The end-to-end test was updated because it now relies on the frontend
mechanism. I also removed a comment which was there to explain that a
test could be removed after ICRC-2 because now the test can be kept.

# Todos

- [x] Add entry to changelog (if necessary).
existing entry updated