Make icpAccountDetailsStore a queued store (#5713)
# Motivation

We've been seeing test flakiness in `e2e/accounts.spec.ts` since
query+update has been rolled out for 10%.
The problem is that a subaccount is created before the original
`queryAccount` update response comes back.
So when it finally comes back with a response from before the subaccount
was created, it's as if the new subaccount suddenly disappears.

We've had similar issues before. The solution is to use `queuedStore`
which is specifically made to make sure old update responses don't
overwrite newer query responses.

# Changes

1. Change `icpAccountDetailsStore` to be a queued store.
2. Update `frontend/src/lib/services/icp-accounts.services.ts` to use
the queued store API when populating `icpAccountDetailsStore`.
3. Add `setForTesting` to `queuedStore` to make it easier to set a value
when one doesn't care about testing race conditions.

# Tests

1. Add a unit test to `frontend/src/tests/lib/routes/Accounts.spec.ts`
which demonstrates the flakiness that we were seeing.
2. Update a unit test in
`frontend/src/tests/lib/services/icp-accounts.services.spec.ts` which
was already testing the same race condition in balances store to also
test the account details store.
3. Ran the e2e test that was flaky 100 times without failure. (I only
ran it up to the point where it was flaking before, to speed things up.)

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary