FOLLOW-244: Claim unclaimed neurons from the frontend (#5734)
# Motivation

Staking a neuron requires 2 steps: transferring the stake and claiming
the neuron.
Both steps are done in ic-js.
But it's possible that the process gets interrupted in between.
In this case the nns-dapp canister finishes the process.
We want to move this fallback mechanism to the frontend.

The functionality was added in previous PRs. Here was call the service
function from the wallet component to make it actually functional.

# Changes

1. Call `claimNeuronsIfNeeded` from `NnsWallet`.

# Tests

1. Tested manually after disabling the functionality in the canister, by
reloading the page as soon as the staking process is started.
2. Unit tests added for main account, subaccount and hardware wallet
account.

# Todos

- [x] Add entry to changelog (if necessary).