Add missing arg-sha256 parameter to release (#5747)
# Motivation

When creating the upgrade proposal for nns-dapp, `ic-admin` requires a
`--arg-sha256` parameter, which is currently missing in the release
script. This omission results in the following error message:

```
thread 'main' panicked at rs/registry/admin/src/main.rs:5472:28:
Must provide a sha256 checksum for the upgrade arg
```

This requirement seems to have been added in
https://github.com/dfinity/ic/pull/1640

# Changes

- Add missing parameter.

# Tests

- Created the latest release using these changes.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.