Add getLoadedSnsAggregatorData (#5765)
# Motivation

We want to create SNS wrappers based on aggregator data to avoid
unnecessary calls to SNS-W and root canisters.

To do this we need to be able to wait for the `snsAggregatorStore` to be
loaded if wrappers are requested before the aggregator data is loaded.

In this PR we add just a function to get the data from the
`snsAggregatorStore` after it is loaded.

# Changes

1. Add `getLoadedSnsAggregatorData` which waits for the
`snsAggregatorStore` to have defined data and then returns that data.

# Tests

1. Unit tests added.
2. Tested manually in another branch.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary