Restore all mocks in NnsProposalDetail.spec.ts (#5769)
# Motivation

To avoid unpredictable behavior, each test should start with a clean
slate.
One component of that is restoring all mocks before each test.
To be able to do this, mock behavior needs to be defined in `beforeEach`
rather than the top-level scope.

This PR does that for `NnsProposalDetail.spec.ts`.

# Changes

1. Mock `proposalsApi.queryProposal` in `beforeEach`.
2. Remove mocking of `markdownToHTML` as it seems to be unnecessary.
3. Add `vi.restoreAllMocks();` in `beforeEach`.

# Tests

Still passes.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary