Add scripts/ic-admin-propose (#5801)
# Motivation

Creating a proposal via `ic-admin` is always tedious because you need to
construct arguments for `--nns-url`, `--secret-key-pem` and
`--proposal`.

For the sake of upgrading a canister, `scripts/propose-install-canister`
does this for you, but for other types of proposals it's still a hassle.

# Changes

1. Extract `scripts/ic-admin-propose` from
`scripts/propose-install-canister` to create the necessary arguments for
arbitrary proposals.
2. Use `scripts/ic-admin-propose` in `scripts/propose-install-canister`.
3. Use `webserver-port` instead of `replica-port`. It does not seem to
make a difference but when switching to PocketIC, `webserver-port` gives
the correct port to use and `replica-port` no longer works.

# Tests 

1. Used `scripts/propose-install-canister` to upgrade nns-dapp locally
with
```
scripts/propose-install-canister --canister nns-dapp --wasm nns-dapp.wasm.gz -a nns-dapp-arg-local.did
```

2. Used `scripts/ic-admin-propose` to create an "Update SNS-W Subnet
Ids" proposal locally with
```
scripts/ic-admin-propose propose-to-update-sns-subnet-ids-in-sns-wasm --sns-subnet-ids-to-add wz6mc-fcmtc-vbptg-sltyl-iljh5-sgwbf-uzlrl-uihqb-huk5d-l476q-gae  --sns-subnet-ids-to-remove ewlco-wjy6h-oafn4-sjmhf-litbm-aout5-atu3t-ygdsc-mqpmz-vpx73-sae
```

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary