Add scripts/dfx-server-origin (#5803)
# Motivation

Some scripts require a URL to the replica that depends on the dfx
network being used.
It's useful to have this in a separate script to be able to reuse it.
I want to use this later to make `scripts/sns/aggregator/get-sns`
support networks other than mainnet.

# Changes

1. Extract the code to get the server URL from
`scripts/ic-admin-propose` and create a new script
`scripts/dfx-server-origin` with it.

# Tests

Manually tried using `propose-install-canister` (which uses
`scripts/ic-admin-propose`) to install nns-dapp on `local` and on
`devenv_dskloet`.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary