Show loaded swap commitments even if some fail to load (#5817)
# Motivation

On the Launchpad we show the user's swap commitments for all swaps in
which they participated.
But if one of the swap canisters ran out of cycles and the commitment
for that swap fails to load, all the projects show spinners even though
some of the commitments might have loaded successfully.

This is because we use `querySnsSwapCommitments` which returns an error
if any of the commitments fails to load and then we reset the entire
`snsSwapCommitmentsStore`.

This is unnecessary. We can still load the successfully loaded
commitments into the store.

# Changes

1. Instead of using `querySnsSwapCommitments`, just call
`loadSnsSwapCommitment` for each SNS.
2. Add a missing `await` in `loadSnsSwapCommitment` which was caught by
the new tests.
3. Remove unused code.

# Tests

1. Tested manually on a local replica which has a frozen swap canister.
2. Existing unit tests adapted to the new behavior.
4. New unit test added where some commitments are loaded

# Todos

- [x] Add entry to changelog (if necessary).