FOLLOW-244: Add notifyTopUpIfNeeded to canisters service (#5857)
# Motivation

Topping up a canister is a 2-step process:

1. Transfer ICP to a subaccount of the CMC
2. Notify the CMC of the transfer
Because the process might be interrupted between the 2 steps, the
nns-dapp canister listens for all transactions to see if any might be a
canister top-up transaction by an NNS dapp user.

We want the NNS dapp to stop listening for all transactions so we want
to move this fallback mechanism to the frontend.

In this PR we add a service function that will be used to check if the
CMC needs to be notified and to do so if needed.

It only checks the most recent transaction to see if it is a funding
transaction. It's theoretically possible to have a funding transaction
that is followed by another transaction that's not the corresponding
burn transaction. This was discussed with the PM and decided that it's
unlikely enough that it's not worth the extra complexity to take into
account. If we decide later that we do want to take this into account we
can still recover from those cases that were missed before.

# Changes

1. Add `notifyTopUpIfNeeded`. It checks the latest transaction on the
CMC subaccount for a specific canister and notifies the CMC if the
latest transaction is a top-up transaction.

# Tests

1. Added unit tests.
2. Tested manually in a branch with end-to-end changes.

# Todos

- [ ] Add entry to changelog (if necessary).
not yet