Keep existing data in SNS aggregator when new data can't be loaded (#5858)
# Motivation

The SNS aggregator queries the root, swap, governance and ledger
canisters of each SNS periodically to store important information about
each SNS so it can be easily fetched with a single call.

For some of these calls (the ones to the swap canister), if the call
fails, the data the aggregator has is replaced with empty default data.

For the calls to the other canisters, it's treated as an error and the
SNS isn't updated at all. Not even with data that could be fetched.

This means that if the swap canister of an SNS runs out of cycles, and
then the aggregator tries to update its data on this SNS, important
information such as the lifecycle of the SNS is lost. This also means
that when the NNS dapp tries to filter out aborted SNSes, it can't know
whether this SNS is aborted and so it gets filtered out and it's as if
this SNS doesn't exist.

To avoid this, we should instead keep existing data when a call to a
canister fails.

# Changes

1. Add a getter to get the existing cached data for an SNS from the
state.
2. When updating the data of an SNS, if a call fails, instead of either
failing the entire update or losing the data, keep the existing data for
that part. This applies to `list_sns_canisters`, `get_metadata`,
`list_nervous_system_functions`, `get_state`, `icrc1_metadata`,
`icrc1_fee`, `icrc1_total_supply`, `get_sale_parameters`, `get_init`,
`get_derived_state`, `get_lifecycle`, and
`get_nervous_system_parameters`.

# Tests

1. Unfortunately the SNS aggregator doesn't have unit tests and
refactoring to make it unit-testable would currently be too much work.
2. Normally we would test this end-to-end running against `dfx start`
but making SNS canisters run out of cycles is not really supported yet,
although it might be supported soon.
3. I've tested this manually as follows:

Using `dfx` built from source in order to be able to use dfx with
PocketIC. I've then used a prototype PocketIC CLI available at branch
`mraszyk/pic-cli` of `dfinity/ic` in order to add myself as a controller
to the SNS canisters. Then I used
```
dfx canister update-settings <canister-id> --freezing-threshold 18446744073709551615 --confirm-very-long-freezing-threshold
```
to make canisters be "out of cycles".

With the changes in this PR, a round of aggregator updates would no
longer cause data loss and the SNS remained visible in NNS dapp.

# Todos

- [x] Add entry to changelog (if necessary).

---------

Co-authored-by: Arshavir Ter-Gabrielyan <arshavir.ter.gabrielyan@dfinity.org>